Models code and DB schema complete. First pass on REST API
All models are now done and the database schema has been finished. REST API is on its way, though the current implementation is quite monolithic. There is a way to make it nicer by having each file within the 'rest' package (name not final) make a urlpatterns list that can be included and mapped to specific elements. Even better would be to have a more well defined interface class in python to which to map the classes can be mapped.