Big Model refactor, cut out a lot of AJAX.
Served website is now baked by Django Python. This makes it a lot easier to handle and avoids me having to deal with clientside AJAX. The "index.html" is now mainly data driven. Sections are still static, but they're content is not.

Makes for better maintenance. No JS anymore. All power to python!